Add workaround to compile with clang
When including LLVM headers and using clang (tested with clang 16 on Windows, i.e. using the MS STL), compilation fails due to deleting incomplete types via std::unique_ptr. This problem is reported in https://github.com/llvm/llvm-project/issues/59784 and a proposed workaround is to compile with C++20 only. We are lucky that the two files that have that issue in our code do actually not use any C++23 features. Hence we reduce the applied C++ standard flags for those two files when compiling with clang.