[clang-tidy] Fix concurrency-mt-unsafe
Call std::quick_exit instead of std::exit. Currently, the program is not multithreaded and hence there should not be a risk in calling std::exit. On the other hand, the server will end up being multi-threaded eventually and we want to be prepared for that. Since that exit calls in question occur before any real work has started, it is probably done before any threads will be started as well. On the other hand, at that state there is nothing that is imprtant to be cleaned-up as well, so just calling quick_exit should be fine here.