Improve handling times in analysis
The start & end times for process and threads are now stored explicitly as std::chrono::nanoseconds that are relative to the sessions start time. The code that creates those times has been simplified to reduce the number of required narrowing casts and to be more explicit about conversions that are performed.

Additionally this includes fixes where the process end-times were set to the session end time without subtracting the session start time (which made them absolute times, while we expect them to be relative to the start time).