Fix handling invalid UTF-8 in image file name of ETL files
The encoding of the "image_filename" data in a ETL process event is unspecified. It definitely works correctly for ASCII, but other than that, it does not seem to have any meaningful data when the original executable name included other Unicode characters. Hence we at least sanitize the string and replace invalid UTF-8 data from it.