Fix issues in DWARF resolver
This fixes finding the correct section for the symbol, by explicitly handling ELF objects. This allows us to compare the relative address to the offsets of the sections in the binary. Before this change, we tried to compare the relative offset to the address of the section, which includes the default image base offset.

Additionally we make sure that we try to look up a binary for each module only once, and create an empty cache entry in case no binary could be found.