Support downloading DWARF symbols from `debuginfod`
To describe how the binary modules should be found it is now possible to provide the DWARF resolver with additional options. These include some local file directories to search for the binary in. A path map allows to map paths from the input-files to paths on the current file-system.

If the files cannot be found locally, we can now provide a build-id to the DWARF symbol resolver and it will try to download the necessary debug symbols from a debuginfod server.