Fix missing escape sequences in jsonrpc error messages
If an error message includes specific characters like double-quotes, these need to be escaped when creating the json string for the RPC string, which was not handled correctly.
To fix this, the error messages is stored in a nlohmann::json object and the dump() method takes care of all the correct escaping.