Suppress warning in LLVM 17 when using MSVC
It warns about a possible loss of data when converting from an int64 to int32. For some reason the /I:external flag does not suppress that error, even though it is in external code...

I am not entirely sure which versions of MSVC emit the warning. Currently, it can be reproduced with version 19.37.32825 and LLVM 17.0.2.