Fix linker issues in coverage pipeline
The version of `ld` in the GitHub actions VMs seems to be old and we need to force it to emit dwarf-4 symbol information.

Additionally, there is a linker issue about missing type info in the LLVM library. Easiest "fix" seems to be to simply build LLVM with RTTI support.