Fix reading perf.data with multiple event sources
Only kernel events have an attribute id attached, but we tried to read/find them for all (including perf) events. This was working in cases where there is only one event source, since we did never really look for the correct id, but simply took the one and only attribute as default. Put with multiple event sources, this was failing.

To fix this, the special kernel event handling is now done a little bit earlier in the event creation code, before we try to find the correct attributes for an event, and we simply skip this step for non kernel events.