Support compressed buffers in ETL files
Currently, we only support buffers that are compressed with the "XPress" compression algorithm. I am not actually very sure whether the detection of the correct compression format is right, but it seems to work on all ETL files I have locally.

In addition to compression support, the storage arrays for the buffer data are now dynamically allocated on the heap, using the correct size, which should improve support for non-default buffer sizes.