Use one-based instead of zero-based line numbers
Changes the reported line numbers for symbols from being zero-based to being one-based.
One-based line numbers are what is returned by the LLVM routines, and just sticking to them removes the need to subtract 1, which can lead to issues when the the line number is 0 (probably an invalid one) and we use unsigned integer types.